+++ prboom2/src/dsda/stretch.c
@@ -167,8 +167,8 @@ static void InitExTextParam(stretch_param_t* offsets, enum patch_translation_e f
 
 void dsda_UpdateExTextOffset(enum patch_translation_e flags, int offset) {
   stretch_params_table[patch_stretch_ex_text][flags].deltay1 +=
-    (ST_SCALED_HEIGHT - ex_text_st_scaled_height) * offset * hud_font.line_height / g_st_height +
-    (hud_font.line_height - exhud_font.line_height) * ex_text_scale_y * (offset > 0 ? 1 : -1);
+    (int)((ST_SCALED_HEIGHT - ex_text_st_scaled_height) * offset * hud_font.line_height / g_st_height +
+    (hud_font.line_height - exhud_font.line_height) * ex_text_scale_y * (offset > 0 ? 1 : -1));
 }
 
 void dsda_ResetExTextOffsets(void) {
