+++ prboom2/src/e6y.c
@@ -585,7 +585,7 @@ void e6y_WriteStats(void)
 
   for(i=0; i<TT_MAX; i++) {
     snprintf(str, sizeof(str), "%d", max.stat[i]);
-    max.stat[i] = strlen(str);
+    max.stat[i] = (int)strlen(str);
   }
 
   for (level=0;level<numlevels;level++)
@@ -644,7 +644,7 @@ int AccelerateMouse(int val)
 int AccelerateAnalog(float val)
 {
   if (!analog_accelfactor)
-    return val;
+    return (int)val;
 
   if (val < 0)
     return -AccelerateAnalog(-val);
@@ -867,7 +867,7 @@ int GetFullPath(const char* FileName, const char* ext, char *Buffer, size_t Buff
       break;
     }
 
-    Result = SearchPath(dir,FileName,ext,BufferLength,Buffer,&p);
+    Result = SearchPath(dir,FileName,ext,(DWORD)BufferLength,Buffer,&p);
     if (Result)
       return Result;
   }
